Add script to start the server
Sets up GEM_PATH and RACK_ENV, cd into the app dir and run rackup.
Inputs do not need to be propogated since this script will set up GEM_PATH for
itself.
GEM_PATH has too much in it, but follows the pattern of other Guix ruby
packages, have started a mailing list thread about it.

Thin is in the inputs, so there will be a store reference to thin and it will be
used as the rackup server unless overridden.