Switch from Redis to Memcached for geoip cache
We are increasingly treating Redis as an "Important" datastore, with backups and
replication, etc.  This data is really just cache data that does not need to
live forever, and currently takes up well over half the space in our Redis.  We
could run a second Redis in LRU mode, but it's just as easy to use something
simpler for this case.

By using the memcached protocol shipped with EventMachine, we avoid adding any
gems.  Since we already depend on the CBOR gem by way of Dhall, we just add that
to the Gemfile to make it clear we are using it directly now.  Using CBOR in the
cache instead of JSON will give us a small performance boost and has no real
downside especially when it comes to cache data.